# swarm: implement smart dialing logic