--- a/config/config.go
+++ b/config/config.go
@@ -123,6 +123,8 @@ type Config struct {
 
 	DisableMetrics       bool
 	PrometheusRegisterer prometheus.Registerer
+
+	NoDelayNetworkDialRanker bool
 }
 
 func (cfg *Config) makeSwarm(eventBus event.Bus, enableMetrics bool) (*swarm.Swarm, error) {
@@ -172,6 +174,9 @@ func (cfg *Config) makeSwarm(eventBus event.Bus, enableMetrics bool) (*swarm.Swa
 	}
 	if cfg.MultiaddrResolver != nil {
 		opts = append(opts, swarm.WithMultiaddrResolver(cfg.MultiaddrResolver))
+	}
+	if cfg.NoDelayNetworkDialRanker {
+		opts = append(opts, swarm.WithNoDialDelay())
 	}
 	if enableMetrics {
 		opts = append(opts,