@sukunrt 2023-04-25
Some results from a 1 hour simultaneous run of kubo on the same machine 

Total dial cancellations: 
old: 4100
new: 1700
<img width="862" alt="Screenshot 2023-04-25 at 7 53 49 PM" src="https://user-images.githubusercontent.com/2179993/234307682-85bc088f-7fd2-4941-998a-158f21bd9c25.png">

kubo2 is old
kubo is new

this is the prometheus query
```sum by (job, transport) (increase(libp2p_swarm_dial_errors_total{error="canceled"}[$__rate_interval]))```