@marten-seemann 2023-04-25
> Total dial cancellations:
> old: 4100
> new: 1700

Impressive numbers! Two questions:
1. Do you have any idea what the reason for the remaining cancelations is?
2. Do you have any numbers on connection establishment latency? How much are we adding?