@sukunrt 2023-04-25
>Do you have any idea what the reason for the remaining cancelations is?

For some reason there are many quic-draft29 cancellations. Nodes are just reporting a lot of quic addresses and not as many quic-v1 addresses. Still debugging what is causing this.  

>Do you have any numbers on connection establishment latency? How much are we adding?

I'll have to measure this. The handshake connection metric currently measures the latency from the time of dialing, so I'll have to instrument this number.