@sukunrt 2023-04-29
@BigLep 
>To put the number of cancellations in context, how many total connections were established during this same 1 hour window?

Sorry, but I somehow managed to delete this prometheus data