@sukunrt 2023-04-29
@marten-seemann 

>Should we prioritize WebTransport over TCP (in cases where we don't have QUIC)?

This is the strategy in the current PR

>Do I understand correctly that we're dialing IPv6 and IPv4 QUIC addresses in parallel?

Yes, I now think we should change this and dial all ipv4 addresses 300ms after ipv6. The PR has this strategy because my isp blocks all ipv6 

>What happens if a node gives us multiple QUIC IP addresses (of the same address family). Should we just randomly pick one and dial it?

Excellent idea. I did some experiments and found that if a peer shares a 4001 port address and another port address, the 4001 address is more likely to be the correct one. So the strategy I've used is to sort the addresses by port number, it is likely that nodes will dial out of a much higher port than the one they choose to listen on. 

Some more numbers: 

kubo on a t2micro aws instance with both ipv4 and ipv6 support. 

happy eyeballs (public == private | quic > tcp | ipv6 > ipv4 ): 

This strategy is essentially what @marten-seemann suggests only difference being that we prioritise ipv6 over ipv4


here we first use quic addresses and then use tcp addresses
within a transport group we rank ipv6 over ipv4  
The first address of the group is dialed immediately and the rest all are dialed after 300ms  
The tcp group is dialed 300ms after the last quic dial  
ex: quic1, quic2, quic3, tcp1, tcp2, tcp3  
quic1: 0, quic2: 300, quic3: 300, tcp1: 600: tcp2: 900, tcp3: 900  
public and private addresses are dialed parallely using the same logic  


PR: (ip4 == ip6 == private | quic > tcp )
        strategy of the pr. all tcp addresses delayed by 300ms

master: no delay

single-dial (public == private | quic > tcp | ipv6 > ipv4): 
        same as happy eyeball but we dial one address at a time and wait 300ms for a result. 

All latency numbers are in milliseconds

Successes is the number of successful outgoing dials which resulted in a connection

| Strategy | Cancellations | Successes | Cancel Fraction | Latency (50p) | Latency (80p) | Latency (90p) | Latency (95p) | 
| :---------- | :-----------: | :---------: | :---------: | :-------------: |:-----------: | :------------: | :---------: |
| master | 1950 | 1600 | 0.54 | 90 | 200 | 240 | 310 |
| happy eyeballs | 510 | 1550 | 0.24 | 94 | 219 | 360 | 650 |
| PR | 1050 | 1997 | 0.34 | 93 | 200 | 270 | 538 |
| single-dial | 520 | 1450 | 0.26 | 95 | 212 | 340 | 600 |


@BigLep

>To put the number of cancellations in context, how many total connections were established during this same 1 hour window?

I'm sorry I somehow deleted prometheus data for that run 🤦‍♂️
But the above numbers are more representative and reproducible. The previous numbers were obtained on my dev machine with a isp that doesn't support ipv6 and I think simultaneous kubo runs aren't very comparable.