@sukunrt 2023-04-29
@marten-seemann 
>Do you have any idea what the reason for the remaining cancelations is?

Some cancellations are because the user is cancelling the dials. No successful connection is made. 
Some cancellations are because tcp dial succeeds and we cancel the quic dial. Some cancellations are because we had multiple quic dials and had to cancel one of them. 

In the graphs below, the first run is master(no delay), the second run is happy-eyeballs, the third run is this pr strategy where all quic addresses are dialed together. 

quic-<nil> means we cancelled a quic dial and there was no successful connection
quic-tcp means we cancelled a quic dial and there was a successful tcp connection. 

<img width="1307" alt="Screenshot 2023-04-30 at 12 20 22 AM" src="https://user-images.githubusercontent.com/2179993/235319369-08812a92-c28b-41c0-a2b5-04d40e377709.png">

Here you can see, there's not much impact on cancellations where there was no successful connection. 

<img width="1270" alt="Screenshot 2023-04-30 at 12 22 16 AM" src="https://user-images.githubusercontent.com/2179993/235319417-3ccbb2c8-0ec0-404c-a0bf-d8fcf9a80a6a.png">

Here we can see that tcp-quic(tcp cancelled, quic succeeded) is reduced considerably for both strategies as expected

<img width="1305" alt="Screenshot 2023-04-30 at 12 23 22 AM" src="https://user-images.githubusercontent.com/2179993/235319448-e6f82c2a-7208-4064-995c-2211463e38ae.png">

The happy eyeballs strategy(middle one) considerably reduces quic-quic and quicv1-quicv1 cancellations

<img width="1305" alt="Screenshot 2023-04-30 at 12 23 22 AM" src="https://user-images.githubusercontent.com/2179993/235319484-47df2841-370c-462d-ad9f-4dfc4512d5e9.png">

None of the strategies have much of an impact in case the successful connection was over tcp. as expected.