@sukunrt 2023-05-09
@marten-seemann 
>I still need to actually understand what dial worker loop is doing. I have to admit I'm pretty lost...

I'll add more comments to the code.