@sukunrt 2023-05-12
>The fix was released in [benbjohnson/clock@v1.3.4 (release)](https://github.com/benbjohnson/clock/releases/tag/v1.3.4).

Continuing the discussion here: 
timer.Reset is not handled in v1.3.4
I've raised https://github.com/benbjohnson/clock/pull/55

We can keep the current implementation for now. I'll change it to use `benbjohnson/clock` when it's merged.